Treat JSON Schema validation errors as errors
If the Swagger document didn't validate against the JSON Schema, we
previously continued to translate the Swagger document to Refract.

This change stops that behaviour and instead stops the translation. The
problems with translating the "invalid" document is that many types in
the Swagger document might not be correct and cause our translation to
fail and throw JS exceptions. For example if a parameter was not an
object but instead none, or an array.